Refactor mappings to constants, add app version to footer, and update work company names
